--- src/main/java/subway/domain/section/SectionToStationId.java
@@ -1,6 +1,0 @@
-package subway.domain.section;
-
-@FunctionalInterface
-public interface SectionToStationId {
-    Long getStationId(Section section);
-}
