+++ src/test/java/subway/controller/LineAcceptanceTest.java
@@ -9,9 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.domain.line.LineRequest;
-import subway.domain.line.LineResponse;
-import subway.domain.station.StationResponse;
+import subway.controller.line.LineRequest;
+import subway.controller.line.LineResponse;
+import subway.controller.station.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
