+++ src/test/java/subway/controller/StationAcceptanceTest.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.domain.station.StationRequest;
-import subway.domain.station.StationResponse;
+import subway.controller.station.StationRequest;
+import subway.controller.station.StationResponse;
 
 import java.util.Arrays;
 import java.util.List;
